backend/requirements.txt
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.4.0
asyncio==3.4.3
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
cymem==2.0.8
debugpy==1.8.0
decorator==4.4.2
deep-translator==1.11.4
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg==1.4
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
greenlet==2.0.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
hyperframe==5.2.0
idna==2.10
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
langcodes==3.3.0
libretranslatepy==2.1.1
llvmlite==0.41.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
simple-websocket==0.10.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
thinc==8.2.1
tiktoken==0.3.3
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wasabi==1.1.2
wcwidth==0.2.7
weasel==0.3.3
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
